backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

# External API clients
openai==1.6.0
elevenlabs==0.2.27
fal-client==0.4.1
requests==2.31.0

# Celery for async tasks (future)
celery==5.3.4
redis==5.0.1

# Video processing (future)
# ffmpeg-python==0.2.0